services/face-recognition/requirements.txt
Flask==2.1.2
face_recognition==1.3.0
psycopg2-binary==2.9.3
